decouple `log` module
